+++ src/main/java/de/dm/intellij/liferay/gradle/LiferayGradlePropertiesDocumentationBundle.java
@@ -1,0 +1,22 @@
+package de.dm.intellij.liferay.gradle;
+
+import com.intellij.AbstractBundle;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+import org.jetbrains.annotations.PropertyKey;
+
+public class LiferayGradlePropertiesDocumentationBundle extends AbstractBundle {
+
+    private static final String PATH_TO_BUNDLE = "LiferayGradlePropertiesDocumentationBundle";
+    private static final AbstractBundle ourInstance = new LiferayGradlePropertiesDocumentationBundle();
+
+    private LiferayGradlePropertiesDocumentationBundle() {
+        super(PATH_TO_BUNDLE);
+    }
+
+    @Nullable
+    public static String message(@NotNull @PropertyKey(resourceBundle = PATH_TO_BUNDLE) String key, @NotNull Object... params) {
+        return ourInstance.messageOrNull(key, params);
+    }
+
+}